refactor to create custom modifier for sunset textfields and securefields.
